portable; The previous commit was a bit overzealous..
The real issue with realloc(3)/reallocarray implementations on
certain systems is that they will free previous memory if called
with a nmemb or size of 0, and return NULL. This has led to
double-free issues in the wild.

On OpenBSD, realloc/reallocarray do *NOT* free the previous
allocation, instead return the equivalent of malloc(0), which
additionally on OpenBSD is protected with PROT_NONE, which means
SIGSEGV on read/write, but it can still be passed to free(3).